fix(IT Wallet): [SIW-1535] Credential duplication after renewal (#6149)
## Short description
This PR fixes a bug that happens when there is a credential in the
wallet and that same credential is renewed. The bug is caused by:
- Always appending new credentials to
`features.itWallet.credentials.credentials` array
- Using the unique key tag as the key for wallet cards

## List of changes proposed in this pull request
- Overwrite credentials of the same type when dispatching
`ITW_CREDENTIALS_STORE`
- `ITW_{credential_type}` is used as key for wallet cards to avoid
duplicates
- Removed `ITW_CREDENTIALS_MULTIPLE_UPDATE` action: now it is safe to
dispatch `ITW_CREDENTIALS_STORE` to store a credential or an array of
credentials and overwrite previous instances

## How to test
Currently one way to test this is to renew the driving license after
it's been revoked, and then navigate to the wallet screen. Without the
fix two driving license cards will appear, with the fix only one.