chore(IT Wallet): [SIW-1518] Add Health Insurance card details screen (#6154)
⚠️ Depends on https://github.com/pagopa/io-app/pull/6139

## Short description
This PR refactors the credential details screen and adds the ability to
render the details of the European Health Insurance card credential.

## List of changes proposed in this pull request
- Added `getThemeColorByCredentialType` utility function that return the
`CredentialTheme` data for each credential type
- Added correct typography to `ItwCredentialCard` component
- Added `content` to `WellKnownClaim` enum: `content` is the claim that
contains the PDF attachment for the European Health Insurance card,
which should not be rendered in the card details screen.
- Replaced `ItwCredentialClaimsList`
with`ItwCredentialPreviewClaimsList` and `ItwPresentationClaimsSection`
for the preview and details screen respectively. This was necessary due
to customizations needed by both screens:
- Added `ItwCredentialPreviewClaimsList` component, which renders the
claims list in the credential preview screen and does not support the
ability to hide claims
- Refactored `ItwPresentationClaimsSection` which now supports the
ability to render special claims (like the Disability Card QR Code) and
has the "hide claims" toggle.
- Refactored `ItwPresentationCredentialDetailScreen` to correctly
display details for the European Health Insurance card credential:
- Added `ItwPresentationDetailsScreenBase` component, which is a wrapper
for the screen and manages the header animation
- Added `ItwPresentationDetailsHeader` component to render the correct
header based on the credential type: it can render a simple header with
a title or a `ItwPresentationCredentialCard` component
- Added `ItwPresentationAdditionalInfoSection` component to render
additional information required by a credential details screen, which
are not part of the credential claims
   - Added `ItwPresentationFiscalCode` component
- Moved logic for credential card rendering in
`ItwPresentationCredentialCard`
- Added tests for new components
- Updated test snapshots

## How to test
Navigate to the details screen of the European Health Insurance card
details screen and check that the header is rendered correctly and the
claims list is not displayed. Only the issuer claim should be visible.

## Preview



https://github.com/user-attachments/assets/548f5639-cab1-410a-acbb-1067e64370af

---------

Co-authored-by: LazyAfternoons <LazyAfternoons@users.noreply.github.com>