chore(IT Wallet): [SIW-1725] Add activation from internal navigation (#6261)
## Short description
This PR adds the ITW activation route for internal navigation, for
example via a message CTA.

## List of changes proposed in this pull request
- Move the ITW linking configuration into its navigator file;
- Add the `discovery/info` route which opens the activation flow. This
route should be available only when the ITW remote flag is true, the
wallet is not in an active state (thus no eID has been obtained yet) and
the user trial system is active.

## How to test
This can be tested with pagopa/io-dev-api-server#420. Open the test
message sent from the dev server and click `ITW Activation`. You can
also test if it stops working by manipulating the selectors of the
aformentioned states.

---------

Co-authored-by: Mario Perrotta <mario.perrotta@pagopa.it>