chore: [IOBP-925] Add amount parameter when requesting list of all payment methods (#6322)
## Short description
This PR adds the amount filter when requesting the list of all payment
methods. This parameter gives the backend to filter every payment method
that cannot handle that amount;

## List of changes proposed in this pull request
- Added the `amount` property to the request of
paymentsGetPaymentMethodsAction setting it equals to the current amount
of the notice to pay;

## How to test
With the app in local env. and Reactotron or proxyman started, check
that the amount is sent through the query parameter in a digit format
(i.e- If the amount provided on the screen is 65,05€ in a digit format,
it should be 6505) when requesting the API
`/ecommerce/io/v2/payment-methods`