fix(IT Wallet): [SIW-1733] Disable wallet activation from async credential link when the user is not enabled (#6276)
## Short description
This PR covers a rare edge case that happens if somebody is able to get
and launch the async credential deep link without a valid wallet. In
this case there is an entry point to the wallet activation flow,
currently not protected.

The async continuation screen has also been made more robust: now it can
only accept a valid `crendentialType` parameter, otherwise it displays a
generic error message.

## List of changes proposed in this pull request
- Disabled linking when ITW is not enabled or the user has not an active
trial subscription.
- Handle unexpected route parameters with a generic error message.
- Added snapshot tests.

## How to test
Reset the wallet, mock a message with the link
`ioit://itw/credential/issuance?credentialType=MDL` and follow it
without an active subscription. You should not be able to start the
wallet activation flow.

---------

Co-authored-by: LazyAfternoons <LazyAfternoons@users.noreply.github.com>
Co-authored-by: LazyAfternoons <lazyafternoons@outlook.it>
Co-authored-by: Mario Perrotta <mario.perrotta@pagopa.it>