fix(IT Wallet): [SIW-1798] Removes unnecessary rendering of failure screens during issuance flows (#6343)
## Short description
This PR addresses an issue where failure screens are rendered with
incorrect props on initial load or, in rare cases, are displayed
unnecessarily. This update ensures that failure screens render only when
appropriate, with the correct failure type.

#### How to reproduce
1. Use a network inspection tool (e.g., Proxyman) to override the
response of either the `/eid` or `/credential` endpoints, forcing them
to return an error status code.
2. Initiate the eID or credential issuance flow.
3. To verify if the failure screens are rendered twice, insert a
`console.log` statement in the `ContentView` component. You should
observe:
   - An initial render with a `GENERIC` failure type.
   - A second render with the actual, specific failure type.

This behavior occurs because the first render processes a `None` failure
type, which defaults to a `GENERIC` failure.

## List of changes proposed in this pull request
- Set `constNull` as the default for failure screens when no failure is
present in the context, removing the `GENERIC` failure fallback.
- Removed generic errors from the preview screen, which were causing
unwanted failure screen renderings during screen transitions.

## How to test
1. Use a network inspection tool (e.g., Proxyman) to override the
response of either the `/eid` or `/credential` endpoints, forcing them
to return an error status code.
2. Initiate the eID or credential issuance flow.
3. Verify if the failure screens are not rendered twice.