fix: [IOBP-950] Filter tabs behavior into biz events notices list (#6363)
## Short description
This PR fixes a strange behavior on the filter tabs in the notices list
screen. When the user taps on a tab and it become selected, it seems
that doesn't stay in "active" mode. This PR fixes it preventing
unnecessary rerendering.

## List of changes proposed in this pull request
- Memoized the `SectionListHeaderTitle` component to avoid unnecessary
re-renders that causes an issue with the component animation of the
TabItem that loses the focus state.

## How to test
- Open the `Pagamenti` section.
- Tap on the CTA `Vedi tutte`.
- Tap on any of the three available tabs in the list and check that the
color is the one that marks the active state.