feat: [IOBP-863,IOBP-864] Removed legacy wallet & payments features (#6210)
## Short description
This PR removes all elements related to the old wallet, payment flow,
and onboarding process. It includes the deletion of screens, components,
sagas, reducers, and other elements associated with these deprecated
features.

**N.B**: The goal of this PR is solely to remove the legacy payments
flow and does not include refactoring the deprecated components
displayed on some screens such as IDPay. Please keep this in mind when
reviewing the PR.

## List of changes proposed in this pull request
- A lot of deletion and code removal ☠️

## How to test
To test this PR, both the new payment flow and the onboarding flow must
work correctly. Additionally, it's important to test a payment initiated
from the Message screen. @forrest57 @Vangaorth, this PR cannot be merged
until it is approved by the communication team.
Please test everything both in local environment and UAT env. (by
enabling the feature flag `Ambiente di test pagoPA` and building the app
with the .env.production variables)

---------

Co-authored-by: Emanuele Dall'Ara <71103219+LeleDallas@users.noreply.github.com>
Co-authored-by: Andrea <andrea.piai@pagopa.it>