fix: [IOCOM-1955] Fix message time going out of screen (#6459)
## Short description
This PR fixes a regression in the MessageListItem that made the
message's received date going out of screen.
|Before|After|
|-|-|
|![Screenshot 2024-11-25 (07 34
49)](https://github.com/user-attachments/assets/1f6c135c-62cc-4bdd-b7df-917af29b4562)|![Screenshot_1732612114](https://github.com/user-attachments/assets/477a1127-c43b-4754-8e48-1334317117b3)|

## List of changes proposed in this pull request
- `flex: 1` was replaced with `flexGrow: 1` (in order to have a more
website-driven styling approach) but this caused the problem described
above. Combining it with `flexShrink: 1` results in having the save
behaviour as `flex: 1`

## How to test
check that the message list is displayed properly, with the date staying
inside the cell.